Fix migration to skip *_archive tables if not exist

+++ database/migrations/2023_07_11_150654_drop_fk_entries_project_id.php
@@ -13,7 +13,7 @@ class DropFkEntriesProjectId extends Migration
      */
     public function up()
     {
-        //drop foreign key constraint on entries table 
+        //drop foreign key constraint on entries table
         Schema::table('entries', function (Blueprint $table) {
             $table->dropForeign('fk_entries_project_id');
         });
@@ -23,23 +23,25 @@ class DropFkEntriesProjectId extends Migration
             $table->dropForeign('fk_branch_entries_project_id');
         });
 
-        //drop foreign key constraint on entries_archive table 
-        Schema::table('entries_archive', function (Blueprint $table) {
-            $table->dropForeign('fk_entries_archive_project_id');
-        });
+        // Drop foreign key constraint on entries_archive table if it exists
+        if (Schema::hasTable('entries_archive')) {
+            Schema::table('entries_archive', function (Blueprint $table) {
+                $table->dropForeign('fk_entries_archive_project_id');
+            });
+        }
 
-        //drop foreign key constraint on branch_entries_archive table 
-        Schema::table('branch_entries_archive', function (Blueprint $table) {
-            $table->dropForeign('fk_branch_entries_archive_project_id');
-        });
+        // Drop foreign key constraint on branch_entries_archive table if it exists
+        if (Schema::hasTable('branch_entries_archive')) {
+            Schema::table('branch_entries_archive', function (Blueprint $table) {
+                $table->dropForeign('fk_branch_entries_archive_project_id');
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         //restore foreign keys constraints
         Schema::table('entries', function (Blueprint $table) {
@@ -49,18 +51,22 @@ class DropFkEntriesProjectId extends Migration
                 ->onDelete('cascade');
         });
 
-        Schema::table('entries_archive', function (Blueprint $table) {
-            $table->foreign('project_id', 'fk_entries_archive_project_id')
-                ->references('id')
-                ->on('projects')
-                ->onDelete('cascade');
-        });
+        if (Schema::hasTable('entries_archive')) {
+            Schema::table('entries_archive', function (Blueprint $table) {
+                $table->foreign('project_id', 'fk_entries_archive_project_id')
+                    ->references('id')
+                    ->on('projects')
+                    ->onDelete('cascade');
+            });
+        }
 
-        Schema::table('branch_entries_archive', function (Blueprint $table) {
-            $table->foreign('project_id', 'fk_branch_entries_archive_project_id')
-                ->references('id')
-                ->on('projects')
-                ->onDelete('cascade');
-        });
+        if (Schema::hasTable('branch_entries_archive')) {
+            Schema::table('branch_entries_archive', function (Blueprint $table) {
+                $table->foreign('project_id', 'fk_branch_entries_archive_project_id')
+                    ->references('id')
+                    ->on('projects')
+                    ->onDelete('cascade');
+            });
+        }
     }
 }
